Support for graphics in the terminal (#5)
* Add Sixel support

Fixes #910

* Implementation of the XTSMGRAPHICS sequence.

* Don't clear cells after the right side of the graphic.

* Don’t erase text behind a sixel image; the image might be transparent

We still add a reference to the graphic in the first cell of every
line under the image, but we don’t erase any of the text in any of the
cells.

* remove an unncessary variable from insert_graphic

* Avoid unnecessary clone when set graphic data.

* Define MAX_GRAPHIC_DIMENSIONS as a 2-elements array.

* support DECSET/DECRST (CSI ? Pm h) to change where the cursor ends up

after printing a sixel image. The default is for the cursor to be
moved to the first column of the line after the image. When we receive
CSI ? 8452 h, we will instead leave the cursor on the last line of the
image, on the next column past the end of the image.

* put the TermMode back to just u32; I had miscounted the bits

* move_forward takes a relative column count, not an absolute column number

* Interprets mode 80 as Sixel Display Mode.

This is reverse of the *sixel scrolling* option, which should match the actual
behaviour of DEC terminals.

For reference: https://github.com/alacritty/alacritty/pull/4763#issuecomment-925374707

* Fill all cells under a graphic with the template.

With the template we can create hyperlinks attached to the graphic.

To avoid reflow issues when a row is shrank, wrapped rows that only contain
graphic cells are discarded. With this approach we loss some info, like the
hyperlink, but the image is always properly positioned in the grid.

* Allow hue values up to 360 in the Sixel parser.

* Allow replacing part of a graphic with text.

When text is added to a cell with a reference to a graphic, an operation is sent
to the OpenGL thread to replace a subregion of the cell with a transparent area.

If the OpenGL driver supports the GL_ARB_clear_texture extension, the region is
updated with glClearTexSubImage. If the extension is not available, the
texture is updated with glTexSubImage2D.

* Highlight graphics to show hints.

Similar to the underline line rendered when the cursor is over an hyperlink, for
graphics we now render a border around the graphic.

* Allow overlapping graphics.

If a graphic is added over another one, the implementation now checks if new
graphic has transparent pixels in every cell. If so, the graphic is appended to
the cell, instead of replacing the previous one.

SmallVec is used to prevent heap allocation when the cell only contains a single
graphic. This should be the most common scenario.

The GPU will store up to 100 textures. If another texture is added when there
are 100, the oldest one is deleted.

* Optimize graphics replacement.

A simple optimization for inserting graphics is to detect when a new graphic is
replacing completely an existing one. If both graphics have the same size, and
the new one is opaque, we can assume that the previous graphic will not be
displayed anymore, so it is not considered when update the graphics list in a
single cell.

This commit also adds serde implementation for GraphicsCell. This was used to
debug the new logic.

* Fix clippy warnings.

* Use hls_to_rgb implementation from libsixel.

* Changes in sixel module to be compatible with oldstable.

- Reimplement abs_diff().
- Use positional arguments to format the error message in assert_color!().

* Initialize cell dimensions when create a Graphics instance.

This fixes a bug that crashes the terminal when a graphic is added before
resizing the window.

* Support GLES2 Renderer in sixel

* Set graphics limit per cell.

The limit per grid is increased to `1000`, and a new limit per cell is added,
set to `20`.

* Add Eq derive to ClearSubregion.

Suggested by clippy.

* CSI XTSMGRAPHICS Pi=2, Pa=1 should return dimensions that fit in text area

* Remove entry about Sixel support in the changelog.

* Update damage when a graphics is inserted.

* Apply rustfmt to term/mod.rs.

* Apply rustfmt and clippy to the current branch.

* Serialize/Deserialize only if the `serde` feature is set.

* Apply clippy suggestions.

* Include Sixel support in Device Attributes response.

The response for `\e[c` (Send Device Attributes) now returns level 2 with the
Sixel extension.

The other extensions are 6 (Selectively Erasable Characters) and 22 (Color Text).
The values are documented in page 04-19 of DEC-STD-070.

* Upgrade vte fork.

With `$ cargo update vte`.

* Use the published vte-graphics crate, instead of a Git repository.

* Fix sixels position when padding>0

---------

Co-authored-by: Ayose <ayosec@gmail.com>
Co-authored-by: Daniel Brooks <db48x@db48x.net>
Co-authored-by: kumattau <kumattau@gmail.com>